requirements.txt
discord.py
Flask
os
random
yt_dlp
asyncio